firstproject/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
boto3==1.42.16
botocore==1.42.16
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.0
cryptography==46.0.3
distro==1.9.0
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==39.0.0
fido2==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
mysqlclient==2.2.7
numpy==2.4.0
openai==2.14.0
packaging==25.0
pillow==12.0.0
pillow_heif==1.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydicom==3.0.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
